cast to float32 and delete default accelerator for Mac

+++ etna/models/nn/deepstate/deepstate.py
@@ -225,11 +225,11 @@ class DeepStateNet(DeepBaseNet):
 
     def make_samples(self, df: pd.DataFrame, encoder_length: int, decoder_length: int) -> Iterator[dict]:
         """Make samples from segment DataFrame."""
-        values_real = df.drop(columns=["target", "segment", "timestamp"] + list(self.embedding_sizes.keys())).values
+        values_real = df.drop(columns=["target", "segment", "timestamp"] + list(self.embedding_sizes.keys())).values.astype(np.float32)
 
         # Categories that were not seen during `fit` will be filled with new category
         for feature in self.embedding_sizes:
-            df[feature] = df[feature].astype(float).fillna(self.embedding_sizes[feature][0])
+            df[feature] = df[feature].astype(np.float32).fillna(self.embedding_sizes[feature][0])
 
         # Columns in `values_categorical` are in the same order as in `embedding_sizes`
         values_categorical = df[self.embedding_sizes.keys()].values.T
